Added support for building tarballs from git.

+++ make-release-tarball.bash
@@ -6,16 +6,16 @@ set -e
 unset GOPATH
 unset GOBIN
 
-# build release tarball from a bzr branch
-DEFAULT_JUJU_CORE="lp:juju-core"
-
+# build release tarball from a bzr or git branch
+DEFAULT_BZR_JUJU_CORE="lp:juju-core"
+DEFAULT_GIT_JUJU_CORE="https://github.com/juju/core.git"
 
 usage() {
-    echo "usage: $0 <BZR_REVNO|GIT_REVISION> [BZR_JUJU_CORE_BRANCH|GIT_REPO]"
+    echo "usage: $0 <BZR_REVNO|GIT_REVISION> [BZR_BRANCH|GIT_REPO]"
     echo "  BZR_REVNO: The juju-core bzr revno to build"
     echo "  GIT_REVISION: The juju-core git revision or branch to build,"
-    echo "  BZR_JUJU_CORE_BRANCH: The juju-core bzr branch; defaults to ${DEFAULT_JUJU_CORE}"
-    echo "  GIT_REPO: The juju-core git repo"
+    echo "  BZR_BRANCH: The juju-core bzr branch; defaults to $DEFAULT_BZR_JUJU_CORE"
+    echo "  GIT_REPO: The juju-core git repo; defaults to $DEFAULT_GIT_JUJU_CORE="
     exit 1
 }
 
@@ -36,27 +36,55 @@ check_deps() {
 
 test $# -ge 1 ||  usage
 check_deps
-REVNO=$1
-JUJU_CORE_BRANCH=${2:-$DEFAULT_JUJU_CORE}
-TMP_DIR=$(mktemp -d --tmpdir=$(pwd))
+
+
+if echo "$1" | grep -E '^[0-9-]*$'; then
+    IS_BZR="true"
+    REVNO=$1
+    JUJU_CORE_BRANCH=${2:-$DEFAULT_BZR_JUJU_CORE}
+    PACKAGE="launchpad.net/juju-core"
+else
+    IS_BZR="false"
+    REVISION=$1
+    JUJU_CORE_REPO=${2:-$DEFAULT_GIT_JUJU_CORE}
+    PACKAGE="github.com/juju/core"
+fi
+
+
+HERE=$(pwd)
+TMP_DIR=$(mktemp -d --tmpdir=$HERE)
 mkdir $TMP_DIR/RELEASE
 WORK=$TMP_DIR/RELEASE
 
-echo "Getting juju-core and all its dependencies."
-GOPATH=$WORK go get -v -d launchpad.net/juju-core/... || \
-    GOPATH=$WORK go get -v -d launchpad.net/juju-core/... || \
-    GOPATH=$WORK go get -v -d launchpad.net/juju-core/...
 
-echo "Setting juju-core tree to $JUJU_CORE_BRANCH $REVNO."
-(cd "${WORK}/src/launchpad.net/juju-core/" &&
- bzr pull --no-aliases --remember --overwrite -r $REVNO $JUJU_CORE_BRANCH)
-
-# Devs moved a package.
-if [[ $JUJU_CORE_BRANCH == 'lp:juju-core/1.18' ]]; then
-    echo "Moving deps to support 1.18 releases."
-    GOPATH=$WORK go get -v github.com/errgo/errgo
-    rm -rf $WORK/src/github.com/juju/errgo
+if [[ $IS_BZR == 'true' ]]; then
+    echo "Getting juju core and all its dependencies."
+    GOPATH=$WORK go get -v -d $PACKAGE/...
+    echo "Setting juju core tree to $JUJU_CORE_BRANCH $REVNO."
+    (cd "$WORK/src/$PACKAGE/" &&
+     bzr pull --no-aliases --remember --overwrite -r $REVNO $JUJU_CORE_BRANCH)
+    # Devs moved a package.
+    if [[ $JUJU_CORE_BRANCH == 'lp:juju-core/1.18' ]]; then
+        echo "Moving deps to support 1.18 releases."
+        GOPATH=$WORK go get -v github.com/errgo/errgo
+        rm -rf $WORK/src/github.com/juju/errgo
+    fi
+else
+    echo "Getting juju core from $JUJU_CORE_REPO."
+    mkdir -p $WORK/src/$PACKAGE
+    git clone $JUJU_CORE_REPO $WORK/src/$PACKAGE
+    echo "Setting juju core tree to $REVISION."
+    cd $WORK/src/$PACKAGE
+    if git ls-remote ./  | grep origin/$REVISION; then
+        git checkout origin/$REVISION
+    else
+        git checkout $REVISION
+    fi
+    echo "Getting juju core's dependencies."
+    GOPATH=$WORK go get -v -d ./...
+    cd $HERE
 fi
+
 
 echo "Updating juju-core dependencies to the required versions."
 GOPATH=$WORK go get -v launchpad.net/godeps
@@ -65,16 +93,16 @@ if [[ ! -f $GODEPS ]]; then
     echo "! Could not install godeps."
     exit 1
 fi
-GOPATH=$WORK $GODEPS -u "${WORK}/src/launchpad.net/juju-core/dependencies.tsv"
+GOPATH=$WORK $GODEPS -u "$WORK/src/$PACKAGE/dependencies.tsv"
 # Remove godeps.
 rm -r $WORK/bin
 
 # Smoke test
-GOPATH=$WORK go build -v launchpad.net/juju-core/...
+GOPATH=$WORK go build -v $PACKAGE/...
 
 # Change the generic release to the proper juju-core version.
 VERSION=$(sed -n 's/^const version = "\(.*\)"/\1/p' \
-    $WORK/src/launchpad.net/juju-core/version/version.go)
+    $WORK/src/$PACKAGE/version/version.go)
 mv $WORK $TMP_DIR/juju-core_${VERSION}/
 
 # Tar it up.
@@ -82,4 +110,4 @@ TARFILE=$(pwd)/juju-core_${VERSION}.tar.gz
 cd $TMP_DIR
 tar cfz $TARFILE --exclude .hg --exclude .git --exclude .bzr juju-core_${VERSION}
 
-echo "release tarball: ${TARFILE}"
+echo "release tarball: $TARFILE"
